Uploaded some progress here
Made the basic loadout for the page, minus the card thumbnails on the right. Will take care if that tomorrow, also I have set up some variables and I produced a mixin for creating a news-card. Working on a small mixin for the smaller news cards on the right hand side of the "home" page. Feel free to tinker/add to any of the scss I have made, and let me know if there is anything you hate about it. I'd still just kind of hammering things down as best I can, so it's not 100% organized.